refactor tags
